Fix error 'Undefined offset: 2'
Problem:
When the client processes a response from BitPay, it errors out with the message: Undefined offset: 2

Cause:
The code is attempting to extract the response status code from raw response.
But the preg_match that is being used is not matching the header.
It is not matching because it is assuming that the beginning would be an HTTP version with one or more decimal places.
e.g. HTTP1.1 will match, but HTTP1, or HTTP2 will not.
The responses are coming with HTTP2 now, so it is no longer matching.

Fix:
Modify the regex to work with HTTP2 also.